samples/xy_web_server_demo/requirements.txt
xy_web_server
xy_request_handler_api
xy_django_model
xy_django_app_information
xy_django_app_api_auth